Here should be size_t.

+++ common/phone/obex/obexgen.c
@@ -470,7 +470,7 @@ static GSM_Error OBEXGEN_ChangePath(GSM_StateMachine *s, char *Name, unsigned ch
 static GSM_Error OBEXGEN_ChangeToFilePath(GSM_StateMachine *s, char *File, bool DirOnly, unsigned char *Buffer)
 {
 	GSM_Error		error;
-	int			Pos;
+	size_t			Pos;
 	unsigned char		*req, req2[200];
 
 	if (Buffer == NULL) {
